requirements/prod_requirements.txt
asgiref==3.9.2
beautifulsoup4==4.14.2
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2.6
fake-useragent==2.2.0
greenlet==3.2.4
idna==3.10
playwright==1.55.0
pyee==13.0.0
requests==2.32.5
soupsieve==2.8
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
gunicorn==23.0.0
djangorestframework==3.16.1